Removed saved() from event creating lifecycle hook
Having played around with it some more, the culprit seems to be in when to use `save()`. It seems that when using `creating`, `save()` should not be called. This, indeed, results in an infinite loop. One could call `save()` but only when listening to `create`. Without using `save()` while listening to `create`, the (meta) attributes aren't saved.

I guess it's indeed better to use `creating` and remove `save()` since that saves a query.